chore: generalize the settings component
